requirements.txt
django
djangorestframework
pillow
face-recognition
opencv-python
numpy
requests
python-dotenv
wheel
